Add files via upload
This is my solution to this Programming Problem ...
Write a program that will take as an input from the user their names, bank account number and bank account balance. Have the user enter at least 5 customer details.
Store the user details in a file called bank_details.txt
Display the stored details on the screen.