Add files via upload
This is a program that I wrote as a solution to the below problem ...

Write a program that will create invoices for hotel customers. The invoices should be stored in a file.
The program should allow the user to indicate if they want to create invoices for more customers after the first entry or not. If the say yes the program should continue if they say no the program should terminate.
The invoice should have the following details
    Name of the customer
    Date the invoice was issued
    List of items ordered and their prices
    Total cost of all the items.
    If the total amount of the items ordered by the customer is above Kshs.5000, the customer should be given 5%discount. The discount given and amount payable by the Customer should also be indicated in the invoice.